+++ ai_readiness/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.1.0"
