+++ pyproject.toml
@@ -1,0 +1,21 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ai-readiness"
+version = "0.1.0"
+description = "AI-Readiness Website Evaluator"
+authors = [{name = "Unknown", email=""}]
+readme = "README.md"
+requires-python = ">=3.8"
+license = {text = "MIT"}
+dependencies = [
+    "streamlit",
+    "requests",
+    "beautifulsoup4",
+    "pandas",
+]
+
+[project.scripts]
+ai-readiness-checker = "ai_readiness.cli:main"
